+++ cpp/libs/include/opendnp3/link/LinkStatistics.h
@@ -21,7 +21,7 @@
 #ifndef OPENDNP3_LINKSTATISTICS_H
 #define OPENDNP3_LINKSTATISTICS_H
 
-#include <cstddef>
+#include <cstdint>
 
 namespace opendnp3
 {
@@ -34,49 +34,49 @@ struct LinkStatistics
 	struct Parser
 	{
 		/// Number of frames discared due to header CRC errors
-		size_t numHeaderCrcError = 0;
+		uint32_t numHeaderCrcError = 0;
 
 		/// Number of frames discared due to body CRC errors
-		size_t numBodyCrcError = 0;
+		uint32_t numBodyCrcError = 0;
 
 		/// Number of frames received
-		size_t numLinkFrameRx = 0;
+		uint32_t numLinkFrameRx = 0;
 
 		/// Number of frames detected with bad / malformed contents
-		size_t numBadLinkFrameRx = 0;
+		uint32_t numBadLinkFrameRx = 0;
 
 		/// number of bad LEN fields received (malformed frame)
-		size_t numBadLength = 0;
+		uint32_t numBadLength = 0;
 
 		/// number of bad function codes (malformed frame)
-		size_t numBadFunctionCode = 0;
+		uint32_t numBadFunctionCode = 0;
 
 		/// number of FCV / function code mismtaches (malformed frame)
-		size_t numBadFCV = 0;
+		uint32_t numBadFCV = 0;
 
 		/// number of frames w/ unexpected FCB bit set (malformed frame)
-		size_t numBadFCB = 0;
+		uint32_t numBadFCB = 0;
 	};
 
 	struct Channel
 	{
 		/// The number of times the channel has successfully opened
-		size_t numOpen = 0;
+		uint32_t numOpen = 0;
 
 		/// The number of times the channel has failed to open
-		size_t numOpenFail = 0;
+		uint32_t numOpenFail = 0;
 
 		/// The number of times the channel has closed either due to user intervention or an error
-		size_t numClose = 0;
+		uint32_t numClose = 0;
 
 		/// The number of bytes received
-		size_t numBytesRx = 0;
+		uint32_t numBytesRx = 0;
 
 		/// The number of bytes transmitted
-		size_t numBytesTx = 0;
+		uint32_t numBytesTx = 0;
 
 		/// Number of frames transmitted
-		size_t numLinkFrameTx = 0;
+		uint32_t numLinkFrameTx = 0;
 	};
 
 	LinkStatistics() = default;
